+++ src/icalendar/cal/journal.py
@@ -150,6 +150,7 @@ class Journal(Component):
     contacts = contacts_property
     status = status_property
     attendees = attendees_property
+    from icalendar.attr import RECURRENCE_ID
 
     @property
     def description(self) -> str:
@@ -191,6 +192,7 @@ class Journal(Component):
         last_modified: date | None = None,
         links: LINKS_TYPE_SETTER = None,
         organizer: vCalAddress | str | None = None,
+        recurrence_id: date | datetime | None = None,
         refids: list[str] | str | None = None,
         related_to: RELATED_TO_TYPE_SETTER = None,
         sequence: int | None = None,
@@ -220,6 +222,7 @@ class Journal(Component):
                 the journal.
             links: The :attr:`~icalendar.Component.links` of the journal.
             organizer: The :attr:`organizer` of the journal.
+            recurrence_id: The :attr:`RECURRENCE_ID` of the journal.
             refids: :attr:`~icalendar.Component.refids` of the journal.
             related_to: :attr:`~icalendar.Component.related_to` of the journal.
             sequence: The :attr:`sequence` of the journal.
@@ -265,6 +268,7 @@ class Journal(Component):
         journal.start = start
         journal.status = status
         journal.attendees = attendees
+        journal.RECURRENCE_ID = recurrence_id
 
         return journal